// src/main/kotlin/org/household/frontend/src/app/models/PowerStorageModel.ts
/**
 * Auto-generated file. Do NOT change, run test 'TypescriptModelGeneratorTest' instead
 */

export class PowerStorageModel {
    smartHomeId: string = '';
    name: string = '';
    currentChargingRate: number = 0;
    maxChargingRate: number = 0;
    maxPower: number = 0;
    currentPower: number = 0;
    enabled: boolean = false;
    id: string = '';
    organisationId: string = '';
    timestampCreated: number = 0;
    timestampLastModified: number = 0;
}